apps/data-pipeline/pyproject.toml
[project]
name = "data-pipeline"
version = "0.1.0"
description = "Filing ETL Pipeline - Extract text from document filings at scale using AWS Batch"
requires-python = ">=3.10"
dependencies = [
    "boto3>=1.34.0",
    "pymupdf>=1.24.0",
    "smart-open[s3]>=7.0.0",
    "beautifulsoup4>=4.12.0",
]

[project.optional-dependencies]
scripts = [
    "psycopg2-binary>=2.9.0",
]
